requirements.txt
streamlit
opencv-python-headless
keras
numpy
Pillow
